refactor: restructure library for a more intuitive module organization
